requirements.txt
streamlit==1.29.0
pymongo==4.6.1
python-dotenv==1.0.0
email-validator==2.1.0
